refactor: database management for production and test
