+++ src/Contexts/Shared/Domain/Database.ts
@@ -1,0 +1,5 @@
+
+export interface Database {
+    connect(): Promise<Object>;
+    close(): Promise<void>;
+}
