+++ src/Contexts/Shared/Infrastructure/DataSources/production.ts
@@ -2,8 +2,6 @@ import * as dotenv from "dotenv";
 import { DataSource } from "typeorm";
 
 dotenv.config();
-import services from "@Api/Services/InfrastructureServices";
-const logger = services.get("services.logger");
 
 // entities
 import { MovementEntity } from "@Movement/Infrastructure/Persistence/MovementEntity";
@@ -19,18 +17,3 @@ export const AppDataSource: DataSource = new DataSource({
         MovementEntity
     ]
 });
-
-
-export const databaseConnect = () => {
-    AppDataSource.initialize()
-    .then(() => {
-        console.log("Database connected");
-        logger.info("Data Source has been initialized!");
-    })
-    .catch((err) => {
-        console.log(err);
-        const msg = `Error trying to connect!`;
-        logger.error(msg);
-        throw new Error(msg);
-    });
-}
