+++ tests/Utils/database.ts
@@ -1,12 +1,15 @@
 import { AppDataSource } from "@Shared/Infrastructure/DataSources/production";
-const timeout = 10;
+import { DatabaseHandler } from "@Shared/Infrastructure/Handlers/DatabaseHandler";
+const timeout = 10000;
+
+const db = new DatabaseHandler(AppDataSource);
 
 export const databaseConnect = async () => {
     jest.setTimeout(timeout);
-    await AppDataSource.initialize();
+    await db.connect();
 };
 
 export const databaseClose = async () => {
     jest.setTimeout(timeout);
-    await AppDataSource.destroy();
+    await db.close();
 };
